xmaker: split files
