+++ pkg/strategy/xmaker/synthetichedge_test.go
@@ -74,192 +74,6 @@ func TestSyntheticHedge_GetQuotePrices_BaseQuote(t *testing.T) {
 	assert.Equal(t, Number(10010*31), ask)
 }
 
-func Test_newHedgeMarket(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	mockCtrl := gomock.NewController(t)
-	defer mockCtrl.Finish()
-
-	market := Market("BTCUSDT")
-	depth := Number(100.0)
-	session, marketDataStream, userDataStream := newMockSession(mockCtrl, ctx, market.Symbol)
-	_ = marketDataStream
-	_ = userDataStream
-
-	doneC := make(chan struct{})
-	hm := newHedgeMarket(&HedgeMarketConfig{
-		Symbol:        "BTCUSDT",
-		HedgeInterval: types.Duration(3 * time.Millisecond),
-		QuotingDepth:  depth,
-	}, session, market)
-	go func() {
-		err := hm.Start(ctx)
-		assert.NoError(t, err)
-		close(doneC)
-	}()
-
-	time.Sleep(10 * time.Millisecond)
-	marketDataStream.EmitConnect()
-	cancel()
-	<-doneC
-
-	assert.NotNil(t, hm)
-	assert.NotNil(t, hm.session)
-	assert.NotNil(t, hm.orderStore)
-}
-
-func TestHedgeMarket_hedge(t *testing.T) {
-	ctx := context.Background()
-	mockCtrl := gomock.NewController(t)
-	defer mockCtrl.Finish()
-
-	market := Market("BTCUSDT")
-
-	session, marketDataStream, userDataStream := newMockSession(mockCtrl, ctx, market.Symbol)
-	_ = userDataStream
-
-	mockExchange := session.Exchange.(*mocks.MockExchange)
-
-	submitOrder := types.SubmitOrder{
-		Market:   market,
-		Symbol:   "BTCUSDT",
-		Quantity: Number(1.0),
-		Side:     types.SideTypeSell,
-		Type:     types.OrderTypeMarket,
-	}
-	createdOrder := types.Order{
-		OrderID:          1,
-		SubmitOrder:      submitOrder,
-		ExecutedQuantity: Number(1.0),
-		Status:           types.OrderStatusFilled,
-	}
-	mockExchange.EXPECT().SubmitOrder(gomock.Any(), submitOrder).Return(&createdOrder, nil)
-
-	depth := Number(100.0)
-	hm := newHedgeMarket(&HedgeMarketConfig{
-		Symbol:        "BTCUSDT",
-		HedgeInterval: types.Duration(3 * time.Millisecond),
-		QuotingDepth:  depth,
-	}, session, market)
-
-	err := hm.stream.Connect(ctx)
-	assert.NoError(t, err)
-
-	marketDataStream.EmitBookSnapshot(types.SliceOrderBook{
-		Symbol: "BTCUSDT",
-		Bids: types.PriceVolumeSlice{
-			{Price: Number(10000), Volume: Number(100)},
-		},
-		Asks: types.PriceVolumeSlice{
-			{Price: Number(10010), Volume: Number(100)},
-		},
-	})
-
-	err = hm.hedge(context.Background(), Number(1.0))
-	assert.NoError(t, err)
-}
-
-func TestHedgeMarket_startAndHedge(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	mockCtrl := gomock.NewController(t)
-	defer mockCtrl.Finish()
-
-	market := Market("BTCUSDT")
-
-	session, marketDataStream, userDataStream := newMockSession(mockCtrl, ctx, market.Symbol)
-	_ = userDataStream
-
-	mockExchange := session.Exchange.(*mocks.MockExchange)
-
-	submitOrder := types.SubmitOrder{
-		Market:   market,
-		Symbol:   "BTCUSDT",
-		Quantity: Number(2.0),
-		Side:     types.SideTypeSell,
-		Type:     types.OrderTypeMarket,
-	}
-	createdOrder := types.Order{
-		OrderID:          1,
-		SubmitOrder:      submitOrder,
-		ExecutedQuantity: Number(2.0),
-		Status:           types.OrderStatusFilled,
-	}
-	mockExchange.EXPECT().SubmitOrder(gomock.Any(), submitOrder).Return(&createdOrder, nil)
-
-	doneC := make(chan struct{})
-	depth := Number(100.0)
-	hm := newHedgeMarket(&HedgeMarketConfig{
-		Symbol:        "BTCUSDT",
-		HedgeInterval: types.Duration(3 * time.Millisecond),
-		QuotingDepth:  depth,
-	}, session, market)
-
-	go func() {
-		err := hm.start(ctx, 40*time.Millisecond)
-		assert.NoError(t, err)
-		close(doneC)
-	}()
-
-	marketDataStream.EmitConnect()
-
-	marketDataStream.EmitBookSnapshot(types.SliceOrderBook{
-		Symbol: "BTCUSDT",
-		Bids: types.PriceVolumeSlice{
-			{Price: Number(10000), Volume: Number(100)},
-		},
-		Asks: types.PriceVolumeSlice{
-			{Price: Number(10010), Volume: Number(100)},
-		},
-	})
-
-	hm.positionDeltaC <- Number(1.0)
-	hm.positionDeltaC <- Number(1.0)
-
-	time.Sleep(100 * time.Millisecond)
-
-	// emit trades
-	userDataStream.EmitTradeUpdate(types.Trade{
-		ID:            1,
-		OrderID:       1,
-		Exchange:      types.ExchangeBinance,
-		Price:         Number(103000.0),
-		Quantity:      Number(1.0),
-		QuoteQuantity: Number(103000.0 * 1.0),
-		Symbol:        "BTCUSDT",
-		Side:          types.SideTypeSell,
-		IsBuyer:       false,
-		Time:          types.Time{},
-		Fee:           fixedpoint.Zero,
-	})
-
-	time.Sleep(100 * time.Millisecond)
-
-	assert.Equal(t, Number(1.0), hm.positionExposure.pending.Get())
-	assert.Equal(t, Number(1.0), hm.positionExposure.net.Get())
-
-	userDataStream.EmitTradeUpdate(types.Trade{
-		ID:            2,
-		OrderID:       1,
-		Exchange:      types.ExchangeBinance,
-		Price:         Number(103000.0),
-		Quantity:      Number(1.0),
-		QuoteQuantity: Number(103000.0 * 1.0),
-		Symbol:        "BTCUSDT",
-		Side:          types.SideTypeSell,
-		IsBuyer:       false,
-		Time:          types.Time{},
-		Fee:           fixedpoint.Zero,
-	})
-
-	time.Sleep(100 * time.Millisecond)
-	assert.Equal(t, Number(0.0), hm.positionExposure.pending.Get())
-	assert.Equal(t, Number(0.0), hm.positionExposure.net.Get())
-
-	cancel()
-	<-doneC
-}
-
 func TestSyntheticHedgeMarket_StartAndHedge(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -448,39 +262,6 @@ func TestSyntheticHedgeMarket_StartAndHedge(t *testing.T) {
 
 	cancel()
 	<-doneC
-}
-
-func TestPositionExposure(t *testing.T) {
-	pe := newPositionExposure("BTCUSDT")
-
-	// initial value
-	assert.Equal(t, Number(0), pe.net.Get())
-	assert.Equal(t, Number(0), pe.pending.Get())
-	assert.Equal(t, Number(0), pe.GetUncovered())
-
-	// open position (maker orders are filled)
-	pe.Open(Number(2))
-	assert.Equal(t, Number(2), pe.net.Get())
-	assert.Equal(t, Number(0), pe.pending.Get())
-	assert.Equal(t, Number(2), pe.GetUncovered())
-
-	// cover 1 (hedge order is placed, and pending is updated)
-	pe.Cover(Number(1))
-	assert.Equal(t, Number(2), pe.net.Get())
-	assert.Equal(t, Number(1), pe.pending.Get())
-	assert.Equal(t, Number(1), pe.GetUncovered())
-
-	// close 1 (hedge order is filled, pending is updated)
-	pe.Close(Number(-1))
-	assert.Equal(t, Number(1), pe.net.Get())
-	assert.Equal(t, Number(0), pe.pending.Get())
-	assert.Equal(t, Number(1), pe.GetUncovered())
-
-	// open 1 (another maker order is filled, pending is updated)
-	pe.Open(Number(-1))
-	assert.Equal(t, Number(0), pe.net.Get())
-	assert.Equal(t, Number(0), pe.pending.Get())
-	assert.Equal(t, Number(0), pe.GetUncovered())
 }
 
 // bindMockMarketDataStream binds default market data stream behaviors
